+++ src/pages/asset/card/CardList.vue
@@ -1,0 +1,21 @@
+<template>
+  <div class="card-list">
+    <CardItem v-for="(card, index) in cards" :key="index" :card="card" />
+  </div>
+</template>
+
+<script setup>
+import CardItem from './CardItem.vue';
+
+const props = defineProps({
+  cards: { type: Array, required: true },
+});
+</script>
+
+<style scoped>
+.card-list {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+</style>
